refactor: backend: remove createImage method, modify updateImage method

+++ backend/app/Data/ProductImageRepository.php
@@ -3,7 +3,6 @@
 namespace App\Data;
 
 use App\Models\Product;
-use Illuminate\Http\File;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
@@ -18,28 +17,12 @@ class ProductImageRepository
         $this->product = new Product();
     }
 
-    public function createImage(Request $request): array
-    {
-        $product = $this->product->find($request->product_id);
-
-        if ($product->image_path != null) {
-            throw new \Exception('image aleady exsists');
-        }
-
-        $this->saveImage($request);
-
-        return [
-            'message' => 'Image uploaded successfully'
-        ];
-    }
-
     public function updateImage(Request $request)
     {
         $product = $this->product->find($request->product_id);
-        $image = $request->file('new_image');
         $image_path = $product->image_path;
         $this->deleteImage($image_path);
-        $this->saveImage($request);
+        $this->saveImage($request->file('new_image'), $request->product_id);
 
         return ['message' => 'product image updated'];
     }
@@ -53,14 +36,13 @@ class ProductImageRepository
         ]);
     }
 
-    private function saveImage(Request $request): void
+    private function saveImage(array|UploadedFile $file, int $id): void
     {
-        $file = $request->file('image');
         $file->store('product');
         $path = Storage::path('product/' . $file->hashName());
         $url = Storage::url('product/' . $file->hashName());
         $this->replaceImage([
-            'product_id' => $request->product_id,
+            'product_id' => $id,
             'new_path' => $path,
             'new_url' => $url
         ]);
@@ -68,6 +50,8 @@ class ProductImageRepository
 
     private function deleteImage($path): void
     {
-        FileFacade::delete($path);
+        if (FileFacade::exists($path)) {
+            FileFacade::delete($path);
+        }
     }
 }
